iLBC: Use uint8_t[] for byte arrays
BUG=909

This is the same as https://review.webrtc.org/41779004/ with the review comments addressed.

R=kwiberg@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/40769004

Cr-Commit-Position: refs/heads/master@{#8394}
git-svn-id: http://webrtc.googlecode.com/svn/trunk@8394 4adac7df-926f-26a2-2b94-8c16560cd09d

+++ webrtc/modules/audio_coding/codecs/ilbc/interface/ilbc.h
@@ -180,21 +180,21 @@ extern "C" {
    *                            -1 - Error
    */
 
-  int16_t WebRtcIlbcfix_Decode(IlbcDecoderInstance *iLBCdec_inst,
-                               const int16_t* encoded,
+  int16_t WebRtcIlbcfix_Decode(IlbcDecoderInstance* iLBCdec_inst,
+                               const uint8_t* encoded,
                                int16_t len,
-                               int16_t *decoded,
-                               int16_t *speechType);
-  int16_t WebRtcIlbcfix_Decode20Ms(IlbcDecoderInstance *iLBCdec_inst,
-                                   const int16_t *encoded,
+                               int16_t* decoded,
+                               int16_t* speechType);
+  int16_t WebRtcIlbcfix_Decode20Ms(IlbcDecoderInstance* iLBCdec_inst,
+                                   const uint8_t* encoded,
                                    int16_t len,
-                                   int16_t *decoded,
-                                   int16_t *speechType);
-  int16_t WebRtcIlbcfix_Decode30Ms(IlbcDecoderInstance *iLBCdec_inst,
-                                   const int16_t *encoded,
+                                   int16_t* decoded,
+                                   int16_t* speechType);
+  int16_t WebRtcIlbcfix_Decode30Ms(IlbcDecoderInstance* iLBCdec_inst,
+                                   const uint8_t* encoded,
                                    int16_t len,
-                                   int16_t *decoded,
-                                   int16_t *speechType);
+                                   int16_t* decoded,
+                                   int16_t* speechType);
 
   /****************************************************************************
    * WebRtcIlbcfix_DecodePlc(...)
